Some test reorg based on code review suggestions
