+++ tests/common/dummy_data.rs
@@ -1,18 +1,14 @@
-use actix_http::StatusCode;
 use actix_web::test::{self, TestRequest};
 use labrinth::{models::projects::Project, models::projects::Version};
 use serde_json::json;
 use sqlx::Executor;
 
-use crate::common::{
-    actix::AppendsMultipart,
-    database::{MOD_USER_PAT, USER_USER_PAT},
-};
+use crate::common::{actix::AppendsMultipart, database::USER_USER_PAT};
 
 use super::{
     actix::{MultipartSegment, MultipartSegmentData},
-    asserts::assert_status,
     environment::TestEnvironment,
+    request_data::get_public_project_creation_data,
 };
 
 #[allow(dead_code)]
@@ -86,12 +82,19 @@ pub async fn add_dummy_data(test_env: &TestEnvironment) -> DummyData {
 }
 
 pub async fn add_project_alpha(test_env: &TestEnvironment) -> (Project, Version) {
-    add_public_dummy_project("alpha", DummyJarFile::DummyProjectAlpha, test_env).await
+    test_env
+        .v2
+        .add_public_project(get_public_project_creation_data(
+            "alpha",
+            DummyJarFile::DummyProjectAlpha,
+        ))
+        .await
 }
 
 pub async fn add_project_beta(test_env: &TestEnvironment) -> (Project, Version) {
     // Adds dummy data to the database with sqlx (projects, versions, threads)
     // Generate test project data.
+    let jar = DummyJarFile::DummyProjectBeta;
     let json_data = json!(
         {
             "title": "Test Project Beta",
@@ -101,7 +104,7 @@ pub async fn add_project_beta(test_env: &TestEnvironment) -> (Project, Version) 
             "client_side": "required",
             "server_side": "optional",
             "initial_versions": [{
-                "file_parts": ["dummy-project-beta.jar"],
+                "file_parts": [jar.filename()],
                 "version_number": "1.2.3",
                 "version_title": "start",
                 "status": "unlisted",
@@ -114,84 +117,6 @@ pub async fn add_project_beta(test_env: &TestEnvironment) -> (Project, Version) 
             }],
             "status": "private",
             "requested_status": "private",
-            "categories": [],
-            "license_id": "MIT"
-        }
-    );
-
-    // Basic json
-    let json_segment = MultipartSegment {
-        name: "data".to_string(),
-        filename: None,
-        content_type: Some("application/json".to_string()),
-        data: MultipartSegmentData::Text(serde_json::to_string(&json_data).unwrap()),
-    };
-
-    // Basic file
-    let file_segment = MultipartSegment {
-        name: "dummy-project-beta.jar".to_string(),
-        filename: Some("dummy-project-beta.jar".to_string()),
-        content_type: Some("application/java-archive".to_string()),
-        data: MultipartSegmentData::Binary(
-            include_bytes!("../../tests/files/dummy-project-beta.jar").to_vec(),
-        ),
-    };
-
-    // Add a project.
-    let req = TestRequest::post()
-        .uri("/v2/project")
-        .append_header(("Authorization", USER_USER_PAT))
-        .set_multipart(vec![json_segment.clone(), file_segment.clone()])
-        .to_request();
-    let resp = test_env.call(req).await;
-
-    assert_eq!(resp.status(), 200);
-
-    // Get project
-    let req = TestRequest::get()
-        .uri("/v2/project/beta")
-        .append_header(("Authorization", USER_USER_PAT))
-        .to_request();
-    let resp = test_env.call(req).await;
-    let project: Project = test::read_body_json(resp).await;
-
-    // Get project's versions
-    let req = TestRequest::get()
-        .uri("/v2/project/beta/version")
-        .append_header(("Authorization", USER_USER_PAT))
-        .to_request();
-    let resp = test_env.call(req).await;
-    let versions: Vec<Version> = test::read_body_json(resp).await;
-    let version = versions.into_iter().next().unwrap();
-
-    (project, version)
-}
-
-pub async fn add_public_dummy_project(
-    slug: &str,
-    jar: DummyJarFile,
-    test_env: &TestEnvironment,
-) -> (Project, Version) {
-    // Adds dummy data to the database with sqlx (projects, versions, threads)
-    // Generate test project data.
-    let json_data = json!(
-        {
-            "title": format!("Test Project {slug}"),
-            "slug": slug,
-            "description": "A dummy project for testing with.",
-            "body": "This project is approved, and versions are listed.",
-            "client_side": "required",
-            "server_side": "optional",
-            "initial_versions": [{
-                "file_parts": [jar.filename()],
-                "version_number": "1.2.3",
-                "version_title": "start",
-                "dependencies": [],
-                "game_versions": ["1.20.1"] ,
-                "release_channel": "release",
-                "loaders": ["fabric"],
-                "featured": true
-            }],
             "categories": [],
             "license_id": "MIT"
         }
@@ -220,24 +145,12 @@ pub async fn add_public_dummy_project(
         .set_multipart(vec![json_segment.clone(), file_segment.clone()])
         .to_request();
     let resp = test_env.call(req).await;
-    assert_status(resp, StatusCode::OK);
 
-    // Approve as a moderator.
-    let req = TestRequest::patch()
-        .uri(&format!("/v2/project/{slug}"))
-        .append_header(("Authorization", MOD_USER_PAT))
-        .set_json(json!(
-            {
-                "status": "approved"
-            }
-        ))
-        .to_request();
-    let resp = test_env.call(req).await;
-    assert_status(resp, StatusCode::NO_CONTENT);
+    assert_eq!(resp.status(), 200);
 
     // Get project
     let req = TestRequest::get()
-        .uri(&format!("/v2/project/{slug}"))
+        .uri("/v2/project/beta")
         .append_header(("Authorization", USER_USER_PAT))
         .to_request();
     let resp = test_env.call(req).await;
@@ -245,7 +158,7 @@ pub async fn add_public_dummy_project(
 
     // Get project's versions
     let req = TestRequest::get()
-        .uri(&format!("/v2/project/{slug}/version"))
+        .uri("/v2/project/beta/version")
         .append_header(("Authorization", USER_USER_PAT))
         .to_request();
     let resp = test_env.call(req).await;
