+++ tests/common/environment.rs
@@ -1,6 +1,9 @@
 #![allow(dead_code)]
 
+use std::rc::Rc;
+
 use super::{
+    api_v2::ApiV2,
     asserts::assert_status,
     database::{TemporaryDatabase, FRIEND_USER_ID, USER_USER_PAT},
     dummy_data,
@@ -9,8 +12,6 @@ use crate::common::setup;
 use actix_http::StatusCode;
 use actix_web::{dev::ServiceResponse, test, App};
 use futures::Future;
-use labrinth::models::{notifications::Notification, projects::Project};
-use serde_json::json;
 
 pub async fn with_test_environment<Fut>(f: impl FnOnce(TestEnvironment) -> Fut)
 where
@@ -29,8 +30,9 @@ where
 // temporary sqlx db like #[sqlx::test] would.
 // Use .call(req) on it directly to make a test call as if test::call_service(req) were being used.
 pub struct TestEnvironment {
-    test_app: Box<dyn LocalService>,
+    test_app: Rc<Box<dyn LocalService>>,
     pub db: TemporaryDatabase,
+    pub v2: ApiV2,
 
     pub dummy: Option<dummy_data::DummyData>,
 }
@@ -47,9 +49,12 @@ impl TestEnvironment {
         let db = TemporaryDatabase::create().await;
         let labrinth_config = setup(&db).await;
         let app = App::new().configure(|cfg| labrinth::app_config(cfg, labrinth_config.clone()));
-        let test_app = test::init_service(app).await;
+        let test_app: Rc<Box<dyn LocalService>> = Rc::new(Box::new(test::init_service(app).await));
         Self {
-            test_app: Box::new(test_app),
+            v2: ApiV2 {
+                test_app: test_app.clone(),
+            },
+            test_app,
             db,
             dummy: None,
         }
@@ -64,6 +69,7 @@ impl TestEnvironment {
 
     pub async fn generate_friend_user_notification(&self) {
         let resp = self
+            .v2
             .add_user_to_team(
                 &self.dummy.as_ref().unwrap().alpha_team_id,
                 FRIEND_USER_ID,
@@ -72,103 +78,9 @@ impl TestEnvironment {
             .await;
         assert_status(resp, StatusCode::NO_CONTENT);
     }
-
-    pub async fn remove_project(&self, project_slug_or_id: &str, pat: &str) -> ServiceResponse {
-        let req = test::TestRequest::delete()
-            .uri(&format!("/v2/project/{project_slug_or_id}"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        let resp = self.call(req).await;
-        assert_eq!(resp.status(), 204);
-        resp
-    }
-
-    pub async fn get_user_projects_deserialized(
-        &self,
-        user_id_or_username: &str,
-        pat: &str,
-    ) -> Vec<Project> {
-        let req = test::TestRequest::get()
-            .uri(&format!("/v2/user/{}/projects", user_id_or_username))
-            .append_header(("Authorization", pat))
-            .to_request();
-        let resp = self.call(req).await;
-        assert_eq!(resp.status(), 200);
-        test::read_body_json(resp).await
-    }
-
-    pub async fn add_user_to_team(
-        &self,
-        team_id: &str,
-        user_id: &str,
-        pat: &str,
-    ) -> ServiceResponse {
-        let req = test::TestRequest::post()
-            .uri(&format!("/v2/team/{team_id}/members"))
-            .append_header(("Authorization", pat))
-            .set_json(json!( {
-                "user_id": user_id
-            }))
-            .to_request();
-        self.call(req).await
-    }
-
-    pub async fn join_team(&self, team_id: &str, pat: &str) -> ServiceResponse {
-        let req = test::TestRequest::post()
-            .uri(&format!("/v2/team/{team_id}/join"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        self.call(req).await
-    }
-
-    pub async fn remove_from_team(
-        &self,
-        team_id: &str,
-        user_id: &str,
-        pat: &str,
-    ) -> ServiceResponse {
-        let req = test::TestRequest::delete()
-            .uri(&format!("/v2/team/{team_id}/members/{user_id}"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        self.call(req).await
-    }
-
-    pub async fn get_user_notifications_deserialized(
-        &self,
-        user_id: &str,
-        pat: &str,
-    ) -> Vec<Notification> {
-        let req = test::TestRequest::get()
-            .uri(&format!("/v2/user/{user_id}/notifications"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        let resp = self.call(req).await;
-        test::read_body_json(resp).await
-    }
-
-    pub async fn mark_notification_read(
-        &self,
-        notification_id: &str,
-        pat: &str,
-    ) -> ServiceResponse {
-        let req = test::TestRequest::patch()
-            .uri(&format!("/v2/notification/{notification_id}"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        self.call(req).await
-    }
-
-    pub async fn delete_notification(&self, notification_id: &str, pat: &str) -> ServiceResponse {
-        let req = test::TestRequest::delete()
-            .uri(&format!("/v2/notification/{notification_id}"))
-            .append_header(("Authorization", pat))
-            .to_request();
-        self.call(req).await
-    }
 }
 
-trait LocalService {
+pub trait LocalService {
     fn call(
         &self,
         req: actix_http::Request,
